backend/requirements.txt
boto3
python-dotenv
mutagen
tinytag
tqdm
requests
